requirements.txt
Django==3.1.7
pyTelegramBotAPI==3.7.6
requests==2.25.1
python-telegram-bot-pagination==0.0.2
